fix: delete hotOnly option
