+++ lib/Server.js
@@ -70,15 +70,10 @@ class Server {
     normalizeOptions(this.compiler, this.options);
 
     // don't move this position because addEntries called by updateCompiler checks this.options.hot|hotOnly
-    this.hot =
-      typeof this.options.hot === 'boolean' ||
-      typeof this.options.hotOnly === 'boolean'
-        ? this.options.hot || this.options.hotOnly
+    this.options.hot =
+      typeof this.options.hot === 'boolean' || this.options.hot === 'only'
+        ? this.options.hot
         : true;
-
-    if (this.hot) {
-      this.options.hot = true;
-    }
 
     updateCompiler(this.compiler, this.options);
 
@@ -742,7 +737,7 @@ class Server {
         this.sockWrite([connection], 'log-level', this.clientLogLevel);
       }
 
-      if (this.hot) {
+      if (this.options.hot) {
         this.sockWrite([connection], 'hot');
       }
