+++ lib/utils/addEntries.js
@@ -43,7 +43,7 @@ function addEntries(config, options, server) {
     /** @type {(string[] | string)} */
     let hotEntry;
 
-    if (options.hotOnly) {
+    if (options.hot === 'only') {
       hotEntry = require.resolve('webpack/hot/only-dev-server');
     } else if (options.hot) {
       hotEntry = require.resolve('webpack/hot/dev-server');
@@ -141,7 +141,7 @@ function addEntries(config, options, server) {
 
       config.entry = prependEntry(config.entry || './src', additionalEntries);
 
-      if (options.hot || options.hotOnly) {
+      if (options.hot || options.hot === 'only') {
         config.plugins = config.plugins || [];
         if (
           !config.plugins.find(
