+++ lib/utils/createConfig.js
@@ -96,12 +96,6 @@ function createConfig(config, argv, { port }) {
   }
 
   // TODO https://github.com/webpack/webpack-dev-server/issues/616 (v4)
-  // We should prefer CLI arg under config, now we always prefer `hotOnly` from `devServer`
-  if (!options.hotOnly) {
-    options.hotOnly = argv.hotOnly;
-  }
-
-  // TODO https://github.com/webpack/webpack-dev-server/issues/616 (v4)
   // We should prefer CLI arg under config, now we always prefer `clientLogLevel` from `devServer`
   if (!options.clientLogLevel && argv.clientLogLevel) {
     options.clientLogLevel = argv.clientLogLevel;
