+++ lib/utils/updateCompiler.js
@@ -57,7 +57,7 @@ function updateCompiler(compiler, options) {
     });
 
     // do not apply the plugin unless it didn't exist before.
-    if (options.hot || options.hotOnly) {
+    if (options.hot || options.hot === 'only') {
       compilersWithoutHMR.forEach((compiler) => {
         // addDevServerEntrypoints above should have added the plugin
         // to the compiler options
