+++ test/helpers/test-server.js
@@ -23,7 +23,6 @@ function startFullSetup(config, options, done) {
   if (
     options.inline === undefined &&
     options.hot === undefined &&
-    options.hotOnly === undefined &&
     options.liveReload === undefined
   ) {
     options.inline = false;
