+++ test/options.test.js
@@ -188,12 +188,8 @@ describe('options', () => {
         failure: [false],
       },
       hot: {
-        success: [true],
-        failure: [''],
-      },
-      hotOnly: {
-        success: [true],
-        failure: [''],
+        success: [true, 'only'],
+        failure: ['', 'foo'],
       },
       http2: {
         success: [true],
